Move some event management RHs to the proper place
Merged events/controllers/management.py into
events/management/controllers.py
