+++ indico/modules/events/blueprint.py
@@ -19,8 +19,6 @@ from __future__ import unicode_literals
 from indico.modules.events.controllers.admin import (RHReferenceTypes, RHCreateReferenceType, RHEditReferenceType,
                                                      RHDeleteReferenceType)
 from indico.modules.events.controllers.display import RHExportEventICAL
-from indico.modules.events.controllers.management import (RHManageReferences, RHManageEventLocation,
-                                                          RHManageEventPersonLinks)
 from indico.web.flask.wrappers import IndicoBlueprint
 
 
@@ -38,9 +36,3 @@ _bp.add_url_rule('/admin/external-id-types/<int:reference_type_id>', 'delete_ref
                  methods=('DELETE',))
 
 _bp.add_url_rule('/event/<confId>/event.ics', 'export_event_ical', RHExportEventICAL)
-_bp.add_url_rule('/event/<confId>/manage/external-ids', 'manage_event_references', RHManageReferences,
-                 methods=('GET', 'POST'))
-_bp.add_url_rule('/event/<confId>/manage/event-location', 'manage_event_location', RHManageEventLocation,
-                 methods=('GET', 'POST'))
-_bp.add_url_rule('/event/<confId>/manage/event-persons', 'manage_event_person_links', RHManageEventPersonLinks,
-                 methods=('GET', 'POST'))
