+++ indico/modules/events/management/controllers.py
@@ -27,17 +27,18 @@ from indico.modules.events import EventLogRealm, EventLogKind
 from indico.modules.events.contributions.models.persons import (ContributionPersonLink, SubContributionPersonLink,
                                                                 AuthorType)
 from indico.modules.events.contributions.models.subcontributions import SubContribution
+from indico.modules.events.forms import EventReferencesForm, EventLocationForm, EventPersonLinkForm
 from indico.modules.events.management.forms import EventProtectionForm
 from indico.modules.events.management.util import can_lock
 from indico.modules.events.management.views import WPEventManagement
-from indico.modules.events.operations import update_event, delete_event
+from indico.modules.events.operations import delete_event, create_event_references, update_event
 from indico.modules.events.sessions import session_settings, COORDINATOR_PRIV_SETTINGS, COORDINATOR_PRIV_TITLES
 from indico.modules.events.util import get_object_from_args, update_object_principals
 from indico.util.i18n import _
-from indico.web.flask.util import url_for, jsonify_data
+from indico.web.flask.templating import get_template_module
+from indico.web.flask.util import url_for
 from indico.web.forms.base import FormDefaults
-from indico.web.util import jsonify_template
-
+from indico.web.util import jsonify_template, jsonify_data, jsonify_form
 from MaKaC.webinterface.rh.base import RH
 from MaKaC.webinterface.rh.conferenceModif import RHConferenceModifBase
 
@@ -216,3 +217,42 @@ class RHMoveEvent(RHConferenceModifBase):
         flash(_('Event "{}" has been moved to category "{}"').format(self.event_new.title, self.target_category.title),
               'success')
         return jsonify_data(flash=False)
+
+
+class RHManageReferences(RHConferenceModifBase):
+    CSRF_ENABLED = True
+
+    def _process(self):
+        form = EventReferencesForm(obj=FormDefaults(references=self.event_new.references))
+        if form.validate_on_submit():
+            create_event_references(event=self.event_new, data=form.data)
+            flash(_('External IDs saved'), 'success')
+            tpl = get_template_module('events/management/_reference_list.html')
+            return jsonify_data(html=tpl.render_event_references_list(self.event_new.references))
+        return jsonify_form(form)
+
+
+class RHManageEventLocation(RHConferenceModifBase):
+    CSRF_ENABLED = True
+
+    def _process(self):
+        form = EventLocationForm(obj=self.event_new)
+        if form.validate_on_submit():
+            update_event(self.event_new, form.data)
+            flash(_('The location for the event has been updated'))
+            tpl = get_template_module('events/management/_event_location.html')
+            return jsonify_data(html=tpl.render_event_location_info(self.event_new.location_data))
+        return jsonify_form(form)
+
+
+class RHManageEventPersonLinks(RHConferenceModifBase):
+    CSRF_ENABLED = True
+
+    def _process(self):
+        form = EventPersonLinkForm(obj=self.event_new, event=self.event_new, event_type=self.event_new.type)
+        if form.validate_on_submit():
+            update_event(self.event_new, form.data)
+            tpl = get_template_module('events/management/_event_person_links.html')
+            return jsonify_data(html=tpl.render_event_person_links(self.event_new.type, self.event_new.person_links))
+        self.commit = False
+        return jsonify_form(form)
