+++ indico/modules/events/templates/management/event_location.html
@@ -4,7 +4,7 @@
 {{ render_event_location_info(event.location_data) }}
 
 (<a href="#" class="js-dialog-action"
-    data-href="{{ url_for('events.manage_event_location', event) }}"
+    data-href="{{ url_for('event_management.manage_event_location', event) }}"
     data-title="{% trans %}Manage location{% endtrans %}"
     data-update="#event-location"
     data-ajax-dialog>
