+++ indico/modules/events/templates/management/event_person_links.html
@@ -8,7 +8,7 @@
 </div>
 
 (<a href="#" class="js-dialog-action"
-    data-href="{{ url_for('events.manage_event_person_links', event) }}"
+    data-href="{{ url_for('event_management.manage_event_person_links', event) }}"
     data-title="{{ title }}"
     data-update="#event-person-links"
     data-ajax-dialog>
