+++ indico/modules/events/templates/management/reference_list.html
@@ -2,7 +2,7 @@
 
 {{ render_event_references_list(references) }}
 (<a href="#" class="js-dialog-action"
-    data-href="{{ url_for('events.manage_event_references', event) }}"
+    data-href="{{ url_for('event_management.manage_event_references', event) }}"
     data-title="{% trans %}Manage external IDs{% endtrans %}"
     data-update="#reference-list"
     data-ajax-dialog>
